requirements.txt
# *
# setuptools
# pip

# --
torch==2.8.0
torchvision==0.23.0
pandas==2.3.2
numpy==2.2.6
pillow==11.3.0

# run.py --show
# tools/qc.py
tk==0.1.0

# tools/ddg.py
duckduckgo_search==8.1.1

# ireqs
insightface==0.7.3
coloredlogs==15.0.1
flatbuffers==25.2.10
humanfriendly==10.0
onnxruntime==1.22.1
open_clip_torch==3.1.0
psycopg2-binary==2.9.10
pgvector==0.4.1

## The following requirements were added by pip freeze:
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
easydict==1.13
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
ftfy==6.3.1
hf-xet==1.1.10
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
opencv-python-headless==4.12.0.88
packaging==25.0
prettytable==3.16.0
primp==0.15.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.0.8
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.9
timm==1.0.19
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
